refactor(auth): modularize token handling by extracting token functions
