+++ backend/src/utils/extractToken.js
@@ -1,0 +1,6 @@
+export const extractToken = (req) => {
+  if (req.cookies?.jwt) {
+    return req.cookies.jwt;
+  }
+  return null;
+};
